requirements.txt
Flask==2.0.3
Flask_Cors==3.0.10
matplotlib==3.5.0
numpy==1.21.2
Pillow==9.1.1
tensorflow
gunicorn
